backend/requirements.txt
bcrypt==4.2.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.0.1
